Quickstart looks great! Added a bit, tried to make the interface docs more clear
I'm writing this on a long plane ride (again lol) so this might mostly be trash, I will try to review this again when I get back to internet and am less of a tired mess. Really great job on the quickstart! I might just leave this for you to do from now on though since you have a batter grasp on rst than me, if you're fine with that, we can chat on discord about it if you want.