Add documentation inline, generate the documentation, host somewhere (#21)
* Need to prod github into letting me make a PR

* Initialized the sphinx documentation, currently ignoring the build directory

* Added documentation for `__init__.py`'s methods

Added some docstrings, setup the rst to autogenerate based on those docstrings

* Prepped the docs files for the other files

Still have to implement autodoc and improve the documentation for these

* Heavily improved the documentation for bot.py, bot.rst bow autogenerates the correct documentation

I also made some very small changes to some other files and to the sphinx config

* Documented collector.py, got intersphinx working

* Added a dev requirements, added a few more pages, added the toctree, 

Also fixed some issues in the docstrings

* Makefile is now system-agnostic (needs active venv)

* slowly and steadily, documentation is coming along

* Removed 4 TODOs, details below

__init__.py ~124: Made issue #23 covering the TODO
__init__.py ~145: Fixed small error in intersphinx link in `collector.py`
bot.py ~72: Made issue #22 covering the TODO
bot.py ~102: Made issue #24, should be ez to fix

* implement timeout in DiscordInteractiveInterface. Just one line.

* Got quickstart done. finally getting the hang of rst syntax

* prettier and clearer.

* Quickstart looks great! Added a bit, tried to make the interface docs more clear

I'm writing this on a long plane ride (again lol) so this might mostly be trash, I will try to review this again when I get back to internet and am less of a tired mess. Really great job on the quickstart! I might just leave this for you to do from now on though since you have a batter grasp on rst than me, if you're fine with that, we can chat on discord about it if you want.

* Working on creating ci testing for the documentation

* Updated travis.yml to include the sphinx test, let's see how this goes

* Fixing issues in travis.yml

* Pulled the remaining todos, they were not really necessary (thanks tired plane me)

* Remove extraneous words and stuff. Let me know if there are any issues with this, if not I'll do the same sort of thing with all/most of the other pages in the coming days.